requirements.txt
yfinance>=0.2.0
newsapi-python>=0.2.7
vaderSentiment>=3.3.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Optional: AI-powered transcript summaries
openai>=1.0.0

# TUI (Terminal User Interface)
textual>=0.40.0
